refactor: restructure and internationalize structured data components
- Move structured data components from shared to page-specific directories
- Add i18n support for all structured data with translations (en, ko, ja, zh)
- Use getServerTranslations for dynamic language loading
- Create separate structured-data files for home, blog, and docs pages
- Add TypeScript types for all structured data messages
- Remove centralized structured-data.tsx in favor of localized components

This improves SEO by providing language-specific structured data for better
search engine understanding and international visibility.
