+++ apps/docs/src/app/[lang]/blog/structured-data.tsx
@@ -1,0 +1,112 @@
+import { getServerTranslations } from "@/i18n/get-server-translations";
+
+export async function BlogListStructuredData({
+  posts,
+  lang,
+}: {
+  posts: Array<{
+    slug: string;
+    title: string;
+    description?: string;
+    date?: string;
+    author?: string;
+    thumbnail?: string;
+  }>;
+  lang: string;
+}) {
+  const t = await getServerTranslations("blogStructuredData", lang);
+  
+  const structuredData = {
+    "@context": "https://schema.org",
+    "@type": "Blog",
+    name: t("name"),
+    description: t("description"),
+    url: `https://ssgoi.dev/${lang}/blog`,
+    inLanguage: lang,
+    publisher: {
+      "@type": "Organization",
+      name: "SSGOI",
+      logo: {
+        "@type": "ImageObject",
+        url: "https://ssgoi.dev/og.png",
+      },
+    },
+    blogPost: posts.map(post => ({
+      "@type": "BlogPosting",
+      headline: post.title,
+      description: post.description,
+      url: `https://ssgoi.dev/${lang}/blog/${post.slug}`,
+      datePublished: post.date,
+      author: {
+        "@type": "Person",
+        name: post.author || "MeurSyphus",
+      },
+      image: post.thumbnail,
+    })),
+  };
+
+  return (
+    <script
+      type="application/ld+json"
+      dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }}
+    />
+  );
+}
+
+export async function BlogPostStructuredData({
+  post,
+  lang,
+}: {
+  post: {
+    slug: string;
+    title: string;
+    description?: string;
+    date?: string;
+    author?: string;
+    thumbnail?: string;
+    tags?: string[];
+  };
+  lang: string;
+}) {
+  const t = await getServerTranslations("blogStructuredData", lang);
+  
+  const structuredData = {
+    "@context": "https://schema.org",
+    "@type": "BlogPosting",
+    headline: post.title,
+    description: post.description || `Read about ${post.title} on ${t("name")}`,
+    url: `https://ssgoi.dev/${lang}/blog/${post.slug}`,
+    datePublished: post.date,
+    dateModified: post.date,
+    author: {
+      "@type": "Person",
+      name: post.author || "MeurSyphus",
+    },
+    publisher: {
+      "@type": "Organization",
+      name: "SSGOI",
+      logo: {
+        "@type": "ImageObject",
+        url: "https://ssgoi.dev/og.png",
+      },
+    },
+    image: post.thumbnail ? {
+      "@type": "ImageObject",
+      url: post.thumbnail,
+    } : undefined,
+    keywords: post.tags?.join(", "),
+    inLanguage: lang,
+    isPartOf: {
+      "@type": "Blog",
+      name: t("name"),
+      url: `https://ssgoi.dev/${lang}/blog`,
+    },
+  };
+
+  return (
+    <script
+      type="application/ld+json"
+      dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }}
+    />
+  );
+}
