--- apps/docs/src/i18n/messages/en.tsx
@@ -1,2 +1,0 @@
-// Re-export from the modular English messages
-export { default } from "./en/index";
