+++ apps/docs/src/i18n/messages/en/blog-structured-data.tsx
@@ -1,0 +1,4 @@
+export const blogStructuredData = {
+  name: "SSGOI Blog",
+  description: "Latest updates, tutorials, and insights about SSGOI page transition library"
+};
