+++ apps/docs/src/i18n/messages/en/docs-structured-data.tsx
@@ -1,0 +1,4 @@
+export const docsStructuredData = {
+  siteName: "SSGOI Documentation",
+  defaultDescription: "Documentation for"
+};
