+++ apps/docs/src/i18n/messages/en/home-structured-data.tsx
@@ -1,0 +1,15 @@
+export const homeStructuredData = {
+  app: {
+    description: "A powerful page transition library that brings native app-like animations to the web. Create smooth, spring-based transitions with state preservation across all frameworks.",
+    keywords: [
+      "page transitions",
+      "animation library", 
+      "react transitions",
+      "vue transitions",
+      "svelte transitions",
+      "spring animations",
+      "view transitions",
+      "web animations"
+    ]
+  }
+};
