--- apps/docs/src/i18n/messages/ja.tsx
@@ -1,2 +1,0 @@
-// Re-export from the modular Japanese messages
-export { default } from "./ja/index";
