+++ apps/docs/src/i18n/messages/ja/blog-structured-data.tsx
@@ -1,0 +1,4 @@
+export const blogStructuredData = {
+  name: "SSGOIブログ",
+  description: "SSGOIページ遷移ライブラリに関する最新のアップデート、チュートリアル、インサイト"
+};
