+++ apps/docs/src/i18n/messages/ja/docs-structured-data.tsx
@@ -1,0 +1,4 @@
+export const docsStructuredData = {
+  siteName: "SSGOIドキュメント",
+  defaultDescription: "ドキュメント："
+};
