+++ apps/docs/src/i18n/messages/ja/home-structured-data.tsx
@@ -1,0 +1,15 @@
+export const homeStructuredData = {
+  app: {
+    description: "ネイティブアプリのようなアニメーションをウェブにもたらす強力なページ遷移ライブラリです。すべてのフレームワークで状態を保持しながら、スムーズでスプリングベースの遷移を作成します。",
+    keywords: [
+      "ページ遷移",
+      "アニメーションライブラリ",
+      "React遷移",
+      "Vue遷移",
+      "Svelte遷移",
+      "スプリングアニメーション",
+      "ビュー遷移",
+      "ウェブアニメーション"
+    ]
+  }
+};
