--- apps/docs/src/i18n/messages/ko.tsx
@@ -1,2 +1,0 @@
-// Re-export from the modular Korean messages
-export { default } from "./ko/index";
