+++ apps/docs/src/i18n/messages/ko/blog-structured-data.tsx
@@ -1,0 +1,4 @@
+export const blogStructuredData = {
+  name: "SSGOI 블로그",
+  description: "SSGOI 페이지 전환 라이브러리에 대한 최신 업데이트, 튜토리얼 및 인사이트"
+};
