+++ apps/docs/src/i18n/messages/ko/docs-structured-data.tsx
@@ -1,0 +1,4 @@
+export const docsStructuredData = {
+  siteName: "SSGOI 문서",
+  defaultDescription: "문서:"
+};
