+++ apps/docs/src/i18n/messages/ko/home-structured-data.tsx
@@ -1,0 +1,15 @@
+export const homeStructuredData = {
+  app: {
+    description: "웹에 네이티브 앱과 같은 애니메이션을 제공하는 강력한 페이지 전환 라이브러리입니다. 모든 프레임워크에서 상태를 유지하면서 부드럽고 스프링 기반의 전환을 만들어보세요.",
+    keywords: [
+      "페이지 전환",
+      "애니메이션 라이브러리",
+      "리액트 전환",
+      "뷰 전환", 
+      "스벨트 전환",
+      "스프링 애니메이션",
+      "뷰 트랜지션",
+      "웹 애니메이션"
+    ]
+  }
+};
