--- apps/docs/src/i18n/messages/types.ts
@@ -1,2 +1,0 @@
-// Re-export from the new types directory
-export * from "./types";
