+++ apps/docs/src/i18n/messages/types/blog-structured-data.ts
@@ -1,0 +1,4 @@
+export type BlogStructuredDataMessages = {
+  name: string;
+  description: string;
+};
