+++ apps/docs/src/i18n/messages/types/docs-structured-data.ts
@@ -1,0 +1,4 @@
+export type DocsStructuredDataMessages = {
+  siteName: string;
+  defaultDescription: string;
+};
