+++ apps/docs/src/i18n/messages/types/home-structured-data.ts
@@ -1,0 +1,6 @@
+export type HomeStructuredDataMessages = {
+  app: {
+    description: string;
+    keywords: string[];
+  };
+};
