+++ apps/docs/src/i18n/messages/types/index.ts
@@ -5,6 +5,9 @@ import { MetadataMessages } from "./metadata";
 import { SidebarMessages } from "./sidebar";
 import { ConsoleMessages } from "./console";
 import { BlogMessages } from "./blog";
+import { HomeStructuredDataMessages } from "./home-structured-data";
+import { BlogStructuredDataMessages } from "./blog-structured-data";
+import { DocsStructuredDataMessages } from "./docs-structured-data";
 
 export type Messages = {
   header: HeaderMessages;
@@ -14,6 +17,9 @@ export type Messages = {
   sidebar: SidebarMessages;
   console: ConsoleMessages;
   blog: BlogMessages;
+  homeStructuredData: HomeStructuredDataMessages;
+  blogStructuredData: BlogStructuredDataMessages;
+  docsStructuredData: DocsStructuredDataMessages;
 };
 
 export type {
@@ -24,4 +30,7 @@ export type {
   SidebarMessages,
   ConsoleMessages,
   BlogMessages,
+  HomeStructuredDataMessages,
+  BlogStructuredDataMessages,
+  DocsStructuredDataMessages,
 };
