--- apps/docs/src/i18n/messages/zh.tsx
@@ -1,2 +1,0 @@
-// Re-export from the modular Chinese messages
-export { default } from "./zh/index";
