+++ apps/docs/src/i18n/messages/zh/blog-structured-data.tsx
@@ -1,0 +1,4 @@
+export const blogStructuredData = {
+  name: "SSGOI博客",
+  description: "关于SSGOI页面过渡库的最新更新、教程和见解"
+};
