+++ apps/docs/src/i18n/messages/zh/docs-structured-data.tsx
@@ -1,0 +1,4 @@
+export const docsStructuredData = {
+  siteName: "SSGOI文档",
+  defaultDescription: "文档："
+};
