+++ apps/docs/src/i18n/messages/zh/home-structured-data.tsx
@@ -1,0 +1,15 @@
+export const homeStructuredData = {
+  app: {
+    description: "一个强大的页面过渡库，为Web带来类似原生应用的动画效果。在所有框架中创建流畅的基于弹簧的过渡效果，同时保持状态。",
+    keywords: [
+      "页面过渡",
+      "动画库",
+      "React过渡",
+      "Vue过渡",
+      "Svelte过渡",
+      "弹簧动画",
+      "视图过渡",
+      "Web动画"
+    ]
+  }
+};
